types: Implement zeek.select_fields(rec, mask)
This is the first shot to select fields (properties) of objects backed by
Zeek records that have a given attribute. Currently supports zeek.ATTR_LOG
only due to its usefulness for the logging use-case.

Usage example:

    zeek.on('SSL::log_ssl', function(rec) {
      const log_rec = zeek.select_fields(rec, zeek.ATTR_LOG);
      zeek.print(`SSL::log_ssl: ${JSON.stringify(log_rec, null, 2)}`);
    });

The attribute mask of `zeek.select_fields()` applies recursively to
nested objects/records.

Also, move all the type conversion code over to Types.{h,cc}.
