+++ dev/nodejs.lsan.supp
@@ -10,6 +10,7 @@ leak:v8::internal::Heap::GenerationalBarrierSlow
 leak:v8::internal::Heap::InsertIntoRememberedSetFromCode
 leak:v8::internal::Heap::PerformGarbageCollection
 leak:v8::internal::JSFunction::EnsureFeedbackVector
+leak:v8::internal::LoadIC::Load
 leak:v8::internal::Object::SetDataProperty
 leak:v8::internal::OptimizedCompilationJob::FinalizeJob
 leak:v8::internal::Runtime_RegExpExec
