+++ src/Nodejs.h
@@ -1,12 +1,14 @@
 #pragma once
 #include <filesystem>
 #include <optional>
+#include <vector>
 
 #include <node/node.h>
 #include <node/v8.h>
 #include <uv.h>
 
 #include "IOLoop.h"
+#include "Types.h"
 #include "ZeekJS.h"
 
 // Prototypes for plugin hooks.
@@ -17,59 +19,6 @@ class Plugin;
 namespace plugin::Nodejs {
 
 class EventHandler;
-
-// Helper class for wrapping a zeek::ValPtr into a v8::Object.
-// creating Zeek
-class ZeekValWrapper {
- public:
-  ZeekValWrapper(v8::Isolate* isolate);
-
-  // Wrap anything into a v8::Value. Some types are converted
-  // directly like strings and numbers. Others return a "proxy"
-  // object which keeps a reference to the original ValPtr and
-  // uses below callbacks.
-  //
-  // Not everything is implemented.
-  v8::Local<v8::Value> Wrap(const zeek::ValPtr& vp);
-
-  struct Result {
-    bool ok;
-    zeek::ValPtr val;
-    std::string error;
-  };
-
-  // Convert a v8::Value to a ValPtr of the given type. If the
-  // conversion fails, e.g. if type is IPAddr, but v8_val not
-  // a string that conforms to an IP, returns ValPtr(nullptr).
-  //
-  // Very little is implemented.
-  Result ToZeekVal(v8::Local<v8::Value> v8_val, const zeek::TypePtr& type);
-
-  // Callbacks used for Zeek sets.
-  static void ZeekRecordEnumerator(const v8::PropertyCallbackInfo<v8::Array>& info);
-  static void ZeekRecordGetter(v8::Local<v8::Name> property,
-                               const v8::PropertyCallbackInfo<v8::Value>& info);
-  static void ZeekRecordQuery(v8::Local<v8::Name> property,
-                              const v8::PropertyCallbackInfo<v8::Integer>& info);
-  // Callbacks used for Zeek tables.
-  static void ZeekTableEnumerator(const v8::PropertyCallbackInfo<v8::Array>& info);
-  static void ZeekTableIndexGetter(uint32_t index,
-                                   const v8::PropertyCallbackInfo<v8::Value>& info);
-  static void ZeekTableGetter(v8::Local<v8::Name> property,
-                              const v8::PropertyCallbackInfo<v8::Value>& info);
-
-  // String conversion helpers
-  v8::Local<v8::String> v8_str_intern(const char* s);
-  v8::Local<v8::String> v8_str(const char* s);
-
- private:
-  v8::Isolate* isolate_;
-  v8::Global<v8::ObjectTemplate> record_template_;
-  v8::Global<v8::ObjectTemplate> table_template_;
-  v8::Global<v8::String> port_str_;
-  v8::Global<v8::String> proto_str_;
-  v8::Global<v8::String> toJSON_str_;
-};
 
 // Class holding Node.js and V8 state.
 class Instance {
@@ -122,12 +71,13 @@ class Instance {
   static void ZeekHookCallback(const v8::FunctionCallbackInfo<v8::Value>& args);
   static void ZeekEventCallback(const v8::FunctionCallbackInfo<v8::Value>& args);
   static void ZeekInvokeCallback(const v8::FunctionCallbackInfo<v8::Value>& args);
+  static void ZeekSelectFieldsCallback(const v8::FunctionCallbackInfo<v8::Value>& args);
 
   v8::Global<v8::Context>& GetContext() { return context_; };
 
   v8::Isolate* GetIsolate() { return isolate_; };
-  [[nodiscard]] v8::Local<v8::Value> Wrap(const zeek::ValPtr& vp) const {
-    return zeek_val_wrapper_.get()->Wrap(vp);
+  v8::Local<v8::Value> Wrap(const zeek::ValPtr& vp, int attr_mask = 0) {
+    return zeek_val_wrapper_.get()->Wrap(vp, attr_mask);
   }
 
   friend class EventHandler;
