+++ tests/Baseline/zeekjs.types.record-attr-log/.stdout
@@ -1,0 +1,11 @@
+### BTest baseline data generated by btest-diff. Do not edit. Use "btest -U/-u" to update. Requires BTest >= 0.63.
+{
+  "count_logged": 4711,
+  "two_fields_log_record": {
+    "f1_logged": "f1",
+    "f2_logged": "f2"
+  },
+  "two_fields_one_log_record": {
+    "f1_logged": "f1"
+  }
+}
